// src/Interfaces/word-model.ts
export interface Word {
  id: string;
  group: number;
  page: number;
  word: string;
  image: string;
  audio: string;
  audioMeaning: string;
  audioExample: string;
  textMeaning: string;
  textExample: string;
  transcription: string;
  textExampleTranslate: string;
  textMeaningTranslate: string;
  wordTranslate: string;
}

export interface UserWord {
  difficulty: 'weak' | 'hard';
  optional?: { word: Word };
  id?: string;
  wordId?: string;
}